FileSystem structure rework and work on small functions
(filesystem.c)
bytes_data is not a memeber of the inode struct anymore.
Instead, we store all the bytes_data inside bytes_array which is a member of FileSystem.
get_FS and put_FS are functions to get and put file system to a file.

(create.c, write.c and ls.c)
Now we can read and write the file system (struct FileSystem) in a file following this order:
1. Write the superblock
2. For each inode, write the inode and then the bytes corresponding to the inode.
create.c is completed, but write.c and ls.c are still in progress

(all)
TODO: Error handling